Create CODE_OF_CONDUCT.md
Working at the Mozilla Global Sprint. One of the parts of making a project more accessible is being transparent about the conduct of contributors. Github provides two templates for a Code of Conduct: the [Contributor Covenant](http://contributor-covenant.org/) and the [Citizen Code of Conduct](http://citizencodeofconduct.org/). The Contributor Covenant is more pertinent to a community of contributors to an open source project. This is also related to JIRA issue [PLFM-4425](https://sagebionetworks.jira.com/browse/PLFM-4425).