[SYNPY-49] Aggregate acl based on groups (#979)
SYNPY-49: Updating the logic when retrieving the ACL for an entity to take into account the public, authenticated, and teams in the system. Any permissions from public are applied, any permissions for authenticated are applied if you are an authenticated user, and any teams you are a part of are taken into account.