Update Dockerfile - Fixes `pandas` installation
# Problem

I realized because of an error with `nf-synapse` that the newer versions of the `synapseclient` docker container do not have `pandas` installed properly.

# Solution

This issue was caused by mixing `apt-get` installing `python3-pandas` with a base Python image. The issue being that Python packages installed via `apt-get` don't automatically integrate with the Python environment from the Docker base image. Fortunately we have a `pandas` extra built into `setup.cfg`, so I switched to installing it that way.

# Testing

Built the container locally, started up a Python terminal inside of it, and ran `import pandas`. If you do this with `sagebionetworks/synapsepythonclient:v4.7.0` you get an import error.